requirements.txt
numpy>=1.24.0
scipy>=1.10.0
mne>=1.5.0
torch>=2.0.0
torch-geometric>=2.4.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
openneuro-py>=2024.1.0
